backend/referrals/urls.py
"""Referral URLs: /api/v1/referrals."""
from django.urls import include, path
from rest_framework.routers import DefaultRouter

from .views import ReferralViewSet

router = DefaultRouter()
router.register("referrals", ReferralViewSet, basename="referral")

urlpatterns = [path("", include(router.urls))]
