backend/pyproject.toml
[tool.poetry]
name = "therapycare-backend"
version = "0.1.0"
description = "TherapyCare API - Django REST Framework backend"
authors = ["TherapyCare Team"]
readme = "README.md"
packages = []
[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0"
djangorestframework = "^3.14"
djangorestframework-simplejwt = "^5.3"
django-cors-headers = "^4.3"
psycopg = {extras = ["binary"], version = "^3.1"}
drf-spectacular = "^0.27"
django-environ = "^0.11"
gunicorn = "^22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"
pytest-cov = "^4.1"
ruff = "^0.4"
black = "^24.0"
# locust = "^2.0"  # optional: pip install locust for load testing

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = ["migrations", "__pycache__", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py312"]
exclude = "migrations"
